// App.Web/src/lib/api/models/MenuCardDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MenuCardDto = {
    id: string;
    categoryKey: string;
    title: string;
    subtitle: string;
    route: string;
    icon?: string;
    requiredPermission?: string;
    featureFlag?: string;
    isEnabled: boolean;
    sortOrder: number;
};